requirements.txt
Django==4.2.19
psycopg2-binary==2.9.9
pandas==2.2.2
requests==2.31.0
tqdm==4.66.4
gunicorn==22.0.0
python-decouple==3.8
Pillow==10.3.0
django-cors-headers==4.3.1
